requirements.txt
tensorflow-gpu==1.15.0
grpcio==1.30.0
grpcio-tools==1.30.0
pybind11==2.6.2
scikit-learn==0.24.1
scipy==1.5.0
pandas
tqdm
h5py
pyyaml
